Move auth utils;
