+++ components/jrg/appwrapper/new/UserMenu.tsx
@@ -22,7 +22,7 @@ import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { cn } from '@/lib/utils';
 import { useTheme } from '@/components/jrg/theme/useTheme';
 import useUser from '@/auth/hooks/useUser';
-import { getGravatarUrl } from '@/lib/gravatar';
+import { getGravatarUrl } from '@/components/jrg/auth/gravatar';
 import { useCompany } from '@/components/interactive/hooks';
 
 type MenuItem = {
