+++ components/jrg/auth/utils.ts
@@ -19,6 +19,26 @@ export const getAuthMode = (): number => {
   }
   return authMode;
 };
+export const generateCookieString = (key: string, value: string, age: string): string =>
+  `${key}=${value}; Domain=${process.env.NEXT_PUBLIC_COOKIE_DOMAIN}; Path=/; Max-Age=${age}; SameSite=strict;`;
+
+export const getQueryParams = (req: NextRequest): any =>
+  req.url.includes('?')
+    ? Object.assign(
+        {},
+        ...req.url
+          .split('?')[1]
+          .split('&')
+          .map((param) => ({ [param.split('=')[0]]: param.split('=')[1] })),
+      )
+    : {};
+
+export const getRequestedURI = (req: NextRequest): string => {
+  console.log(`Processing: ${req.url}`);
+  return req.url
+    .split('?')[0]
+    .replace(/localhost:\d+/, (process.env.APP_URI || '').replace('https://', '').replace('http://', ''));
+};
 
 export const getJWT = (req: NextRequest) => {
   const rawJWT = req.cookies.get('jwt')?.value;
