+++ lib/utils.ts
@@ -1,27 +1,6 @@
 import { clsx, type ClassValue } from 'clsx';
-import { NextRequest } from 'next/server';
 import { twMerge } from 'tailwind-merge';
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
-export const generateCookieString = (key: string, value: string, age: string): string =>
-  `${key}=${value}; Domain=${process.env.NEXT_PUBLIC_COOKIE_DOMAIN}; Path=/; Max-Age=${age}; SameSite=strict;`;
-
-export const getQueryParams = (req: NextRequest): any =>
-  req.url.includes('?')
-    ? Object.assign(
-        {},
-        ...req.url
-          .split('?')[1]
-          .split('&')
-          .map((param) => ({ [param.split('=')[0]]: param.split('=')[1] })),
-      )
-    : {};
-
-export const getRequestedURI = (req: NextRequest): string => {
-  console.log(`Processing: ${req.url}`);
-  return req.url
-    .split('?')[0]
-    .replace(/localhost:\d+/, (process.env.APP_URI || '').replace('https://', '').replace('http://', ''));
-};
